+++ eelbrain/_utils/__init__.py
@@ -1,5 +1,5 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 from .basic import (
-    WrappedFormater, deprecated, intervals, LazyProperty,
+    WrappedFormater, deprecated, deprecated_attribute, intervals, LazyProperty,
     keydefaultdict, n_decimals, natsorted, log_level, set_log_level)
 from .system import caffeine